requirements.txt
requests>=2.31.0
types-requests>=2.31.0
rich>=13.7.0
urllib3>=2.2.0 
pyyaml>=6.0.1
mypy>=1.8.0